Rename parallel module to macros, keep private
