+++ crates/imgal/src/lib.rs
@@ -22,7 +22,7 @@ pub mod image;
 pub mod integration;
 pub mod kernel;
 pub mod overlay;
-mod parallel;
+mod macros;
 pub mod parameter;
 pub mod phasor;
 pub mod simulation;
